pip_requirements_gpu.txt
absl-py==0.9.0
astor==0.8.1
certifi==2019.11.28
cycler==0.10.0
decorator==4.4.1
efficientnet==1.0.0
gast==0.2.2
google-pasta==0.1.8
graphviz==0.13.2
grpcio==1.26.0
h5py==2.10.0
imageio==2.6.1
joblib==0.14.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
Markdown==3.1.1
matplotlib==3.1.2
networkx==2.4
numpy==1.18.1
opencv-python==4.1.2.30
opt-einsum==3.1.0
Pillow==7.0.0
protobuf==3.11.2
pydot==1.4.1
pyparsing==2.4.6
python-dateutil==2.8.1
python-gflags==3.1.2
PyWavelets==1.1.1
scikit-image==0.16.2
scikit-learn==0.22.1
scipy==1.4.1
simpy==3.0.11
simpy.io==0.2.3
six==1.14.0
sklearn==0.0
tb-nightly==1.15.0a20190806
tensorflow-gpu==2.0.0rc1
termcolor==1.1.0
tf-estimator-nightly==1.14.0.dev2019080601
Werkzeug==0.16.0
wincertstore==0.2
wrapt==1.11.2
